requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
streamlit==1.48.0
langchain==0.3.7
langchain-community==0.3.7
langchain-openai==0.2.11
pinecone-client==2.2.4
youtube-transcript-api==0.6.0
python-dotenv==1.1.1
pydantic==2.11.7
pytube==15.0.0
yt-dlp==2025.8.27
openai-whisper==20250625
faiss-cpu==1.12.0
